Add a workaround for the official Android client
Since the latest release of the official WhatsApp client for Android,
random crashes occur when someone using that client starts typing a
message to a whatsapp-purple user. This seems to be related to the way
we reply to typing notices. As a workaround, disable responses to this
type of message for now.

See issue #61 for details.