Only call purple_connection_update_progress() during login
This moves that switch block to a separate function,
waprpl_update_progress(), and calls it only if !wconn->connected.

I did it that way instead of increasing the indentation level just to
make the diff easier to understand - not quite sure if I succeeded at
doing that.

The main purpose of this change is to reduce the verbosity of
whatsapp-purple in bitlbee, which shows "Connection established" for
every single incoming packet as if it was connection progress.